+++ KB.py
@@ -1,0 +1,23 @@
+from utils import first
+
+
+class KB:
+    def __init__(self, sentence=None):
+        if sentence:
+            self.tell(sentence)
+
+    def tell(self, sentence):
+        """Add the sentence to the KB."""
+        raise NotImplementedError
+
+    def ask(self, query):
+        """Return a substitution that makes the query true, or, failing that, return False."""
+        return first(self.ask_generator(query), default=False)
+
+    def ask_generator(self, query):
+        """Yield all the substitutions that make query true."""
+        raise NotImplementedError
+
+    def retract(self, sentence):
+        """Remove sentence from the KB."""
+        raise NotImplementedError
