+++ PropKB.py
@@ -1,0 +1,32 @@
+from KB import *
+from entail_engine import *
+from cnf import *
+
+
+class PropKB(KB):
+    """A KB for propositional logic. Inefficient, with no indexing."""
+
+    def __init__(self, sentence=None):
+        super().__init__(sentence)
+        self.clauses = []
+
+    def tell(self, sentence):
+        """Add the sentence's clauses to the KB."""
+        self.clauses.extend(conjuncts(to_cnf(sentence)))
+
+    def ask_generator(self, query):
+        """Yield the empty substitution {} if KB entails query; else no results."""
+        if tt_entails(Expr('&', *self.clauses), query):
+            yield {}
+
+    def ask_if_true(self, query):
+        """Return True if the KB entails query, else return False."""
+        for _ in self.ask_generator(query):
+            return True
+        return False
+
+    def retract(self, sentence):
+        """Remove the sentence's clauses from the KB."""
+        for c in conjuncts(to_cnf(sentence)):
+            if c in self.clauses:
+                self.clauses.remove(c)
