+++ cnf.py
@@ -103,3 +103,19 @@ def dissociate(op, args):
 
     collect(args)
     return result
+
+
+def is_definite_clause(s):
+    """Returns True for exprs s of the form A & B & ... & C ==> D,
+    where all literals are positive. In clause form, this is
+    ~A | ~B | ... | ~C | D, where exactly one clause is positive.
+    >>> is_definite_clause(expr('Farmer(Mac)'))
+    True
+    """
+    if is_symbol(s.op):
+        return True
+    elif s.op == '==>':
+        antecedent, consequent = s.args
+        return is_symbol(consequent.op) and all(is_symbol(arg.op) for arg in conjuncts(antecedent))
+    else:
+        return False
