+++ expr.py
@@ -142,3 +142,27 @@ def get_symbols(x: 'Expr') -> set:
         return {x}
     else:
         return {symbol for arg in x.args for symbol in get_symbols(arg)}
+
+
+def is_prop_symbol(s):
+    """A proposition logic symbol is an initial-uppercase string.
+    >>> is_prop_symbol('exe')
+    False
+    """
+    return is_symbol(s) and s[0].isupper()
+
+
+def is_definite_clause(s):
+    """Returns True for exprs s of the form A & B & ... & C ==> D,
+    where all literals are positive. In clause form, this is
+    ~A | ~B | ... | ~C | D, where exactly one clause is positive.
+    >>> is_definite_clause(expr('Farmer(Mac)'))
+    True
+    """
+    if is_symbol(s.op):
+        return True
+    elif s.op == '==>':
+        antecedent, consequent = s.args
+        return is_symbol(consequent.op) and all(is_symbol(arg.op) for arg in conjuncts(antecedent))
+    else:
+        return False
